finish setup.sh - attempt 2
formatting the script output

+++ setup.sh
@@ -6,17 +6,18 @@ exit 1
 # Variables
 dhcpcd="dhcpcd.conf"
 hostapd="hostapd.conf"
-networkNumber=""
 
+# Network Number
+networkNumber=""
 read -p "The name of the wifi network and password will end in a number.
 Enter network number: " networkNumber
-
 if [[ -n ${networkNumber//[0-9]/} ]] ;
 then
     echo "Not a number. Exiting..."
     exit 1
 fi
 
+# Install Software Needed
 cd /home/pi
 
 sudo apt update
@@ -35,17 +36,20 @@ if ! grep -Fxq "nohook" $dhcpcd; then
     # \n -> new line
     # \t -> tab
    printf "\ninterface wlan\n\t0static ip_address=192.168.4.1/24\n\tnohook wpa_supplicant" >> $dhcpcd
-else
-    echo "$dhcpcd already setup. Skipping..."
 fi
 
 # https://www.raspberrypi.com/documentation/computers/configuration.html#wifi-cc-rfkill
 sudo rfkill unblock wlan
 
 # write the following to /etc/hostapd/hostapd.conf
-$ssid="RaspberryPi$networkNumber"
-$password="raspberrypi$networkNumber"
-rm $hostapd
+ssid="RaspberryPi$networkNumber"
+password="raspberrypi$networkNumber"
+
+if [[ -d $hostpad ]]; then
+    rm $hostapd
+fi
+
+touch $hostapd
 echo "country_code=US
 interface=wlan0
 ssid=$ssid
@@ -58,12 +62,17 @@ wpa=2
 wpa_passphrase=$password
 wpa_key_mgmt=WPA-PSK
 wpa_pairwise=TKIP
-rsn_pairwise=CCMP"
+rsn_pairwise=CCMP" > $hostapd
 
 # Print out wifi network and password
-echo "Setup completed!
+YELLOW='\033[0;33m'
+GREEN='\033[0;32m'
+RESET='\033[0m'
+clear
+echo -e "Setup completed!
 
-SSID:     $ssid
-PASSWORD: $password
+${YELLOW}SSID:${YELLOW}     ${GREEN}$ssid${RESET}
+${YELLOW}PASSWORD:${YELLOW} ${GREEN}$password${RESET}
 
-Please reboot to enable all services. Please note your wifi network name and password as your SSH connection will be lost and this access point should start its own. `sudo reboot now`."
+⚠️  Please reboot to enable all services. Please note your wifi network name and password as 
+   your SSH connection will be lost and this access point should start its own."
